fixed server trying to load client classes
