+++ src/main/java/io/github/fabricators_of_create/porting_lib/features/MinecraftClientUtil.java
@@ -4,20 +4,17 @@ import io.github.fabricators_of_create.porting_lib.mixins.client.MinecraftAccess
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.resources.language.LanguageManager;
-
-import java.util.Locale;
 
 @Environment(EnvType.CLIENT)
 public final class MinecraftClientUtil {
-	public static float getRenderPartialTicksPaused(Minecraft minecraft) {
-		return get(minecraft).port_lib$pausePartialTick();
-	}
+    private MinecraftClientUtil() {
+    }
 
+    public static float getRenderPartialTicksPaused(Minecraft minecraft) {
+        return get(minecraft).port_lib$pausePartialTick();
+    }
 
-	private static MinecraftAccessor get(Minecraft minecraft) {
-		return (MinecraftAccessor) minecraft;
-	}
-
-	private MinecraftClientUtil() {}
+    private static MinecraftAccessor get(Minecraft minecraft) {
+        return (MinecraftAccessor) minecraft;
+    }
 }
