+++ src/main/java/io/github/fabricators_of_create/porting_lib/mixins/common/PersistentEntitySectionManagerMixin.java
@@ -11,9 +11,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(PersistentEntitySectionManager.class)
 public class PersistentEntitySectionManagerMixin<T extends EntityAccess> {
-	@Inject(method = "addEntity", at = @At("HEAD"), cancellable = true)
-	public void port_lib$addEntityEvent(T entityAccess, boolean loadedFromDisk, CallbackInfoReturnable<Boolean> cir) {
-		if (entityAccess instanceof Entity entity && EntityEvents.ON_JOIN_WORLD.invoker().onJoinWorld(entity, entity.level(), loadedFromDisk))
-			cir.setReturnValue(false);
-	}
+    @Inject(method = "addEntity", at = @At("HEAD"), cancellable = true)
+    public void port_lib$addEntityEvent(T entityAccess, boolean loadedFromDisk, CallbackInfoReturnable<Boolean> cir) {
+        if (entityAccess instanceof Entity entity && EntityEvents.ON_JOIN_WORLD.invoker().onJoinWorld(entity, entity.level(), loadedFromDisk))
+            cir.setReturnValue(false);
+    }
 }
