+++ src/main/java/io/github/fabricators_of_create/porting_lib/mixins/common/ServerLevelMixin.java
@@ -17,20 +17,20 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(ServerLevel.class)
 public abstract class ServerLevelMixin implements LevelExtensions {
 
-	@ModifyReturnValue(method = "getEntityOrPart", at = @At("RETURN"))
-	public Entity port_lib$getMultipart(Entity entity, int id) {
-		if (entity == null) {
-			Int2ObjectMap<PartEntity<?>> partEntityMap = getPartEntityMap();
-			if (partEntityMap != null) {
-				return partEntityMap.get(id);
-			}
-		}
-		return entity;
-	}
+    @ModifyReturnValue(method = "getEntityOrPart", at = @At("RETURN"))
+    public Entity port_lib$getMultipart(Entity entity, int id) {
+        if (entity == null) {
+            Int2ObjectMap<PartEntity<?>> partEntityMap = getPartEntityMap();
+            if (partEntityMap != null) {
+                return partEntityMap.get(id);
+            }
+        }
+        return entity;
+    }
 
-	@Inject(method = "addPlayer", at = @At("HEAD"), cancellable = true)
-	public void port_lib$addEntityEvent(ServerPlayer serverPlayer, CallbackInfo ci) {
-		if (EntityEvents.ON_JOIN_WORLD.invoker().onJoinWorld(serverPlayer, (Level) (Object) this, false))
-			ci.cancel();
-	}
+    @Inject(method = "addPlayer", at = @At("HEAD"), cancellable = true)
+    public void port_lib$addEntityEvent(ServerPlayer serverPlayer, CallbackInfo ci) {
+        if (EntityEvents.ON_JOIN_WORLD.invoker().onJoinWorld(serverPlayer, (Level) (Object) this, false))
+            ci.cancel();
+    }
 }
