+++ src/main/java/team/creative/cmdcam/common/scene/mode/CamMode.java
@@ -1,5 +1,6 @@
 package team.creative.cmdcam.common.scene.mode;
 
+import net.fabricmc.api.EnvType;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.neoforged.api.distmarker.Dist;
@@ -8,6 +9,7 @@ import team.creative.cmdcam.client.CamEventHandlerClient;
 import team.creative.cmdcam.common.math.point.CamPoint;
 import team.creative.cmdcam.common.scene.CamScene;
 import team.creative.cmdcam.common.scene.run.CamRun;
+import team.creative.cmdcam.common.utils.EnvExecutor;
 import team.creative.creativecore.common.util.math.vec.Vec3d;
 import team.creative.creativecore.common.util.registry.NamedTypeRegistry;
 
@@ -31,8 +33,10 @@ public abstract class CamMode {
     
     @OnlyIn(Dist.CLIENT)
     public void finished(CamRun run) {
-        CamEventHandlerClient.resetFOV();
-        CamEventHandlerClient.resetRoll();
+        EnvExecutor.safeRunWhenOn(EnvType.CLIENT, () -> () -> {
+            CamEventHandlerClient.resetFOV();
+            CamEventHandlerClient.resetRoll();
+        });
     }
     
     @OnlyIn(Dist.CLIENT)
@@ -40,17 +44,19 @@ public abstract class CamMode {
     
     @OnlyIn(Dist.CLIENT)
     public void process(CamPoint point) {
-        CamEventHandlerClient.roll((float) point.roll);
-        CamEventHandlerClient.fov(point.zoom);
-        
-        Entity camera = getCamera();
-        if (camera instanceof Player)
-            ((Player) camera).getAbilities().flying = true;
-        
-        camera.absMoveTo(point.x, point.y - camera.getEyeHeight(), point.z, (float) point.rotationYaw, (float) point.rotationPitch);
-        camera.yRotO = (float) point.rotationYaw;
-        camera.xRotO = (float) point.rotationPitch;
-        camera.moveTo(point.x, point.y - camera.getEyeHeight(), point.z, (float) point.rotationYaw, (float) point.rotationPitch);
+        EnvExecutor.safeRunWhenOn(EnvType.CLIENT, () -> () -> {
+            CamEventHandlerClient.roll((float) point.roll);
+            CamEventHandlerClient.fov(point.zoom);
+
+            Entity camera = getCamera();
+            if (camera instanceof Player)
+                ((Player) camera).getAbilities().flying = true;
+
+            camera.absMoveTo(point.x, point.y - camera.getEyeHeight(), point.z, (float) point.rotationYaw, (float) point.rotationPitch);
+            camera.yRotO = (float) point.rotationYaw;
+            camera.xRotO = (float) point.rotationPitch;
+            camera.moveTo(point.x, point.y - camera.getEyeHeight(), point.z, (float) point.rotationYaw, (float) point.rotationPitch);
+        });
     }
     
     public abstract boolean outside();
