+++ src/main/java/team/creative/cmdcam/common/scene/mode/DefaultMode.java
@@ -1,5 +1,6 @@
 package team.creative.cmdcam.common.scene.mode;
 
+import net.fabricmc.api.EnvType;
 import net.minecraft.client.Minecraft;
 import net.minecraft.world.entity.Entity;
 import net.neoforged.api.distmarker.Dist;
@@ -8,6 +9,7 @@ import team.creative.cmdcam.common.math.point.CamPoint;
 import team.creative.cmdcam.common.scene.CamScene;
 import team.creative.cmdcam.common.scene.run.CamRun;
 import team.creative.cmdcam.common.target.CamTarget.SelfTarget;
+import team.creative.cmdcam.common.utils.EnvExecutor;
 import team.creative.creativecore.common.util.math.vec.Vec3d;
 
 public class DefaultMode extends CamMode {
@@ -24,28 +26,34 @@ public class DefaultMode extends CamMode {
     @OnlyIn(Dist.CLIENT)
     public void process(CamPoint point) {
         super.process(point);
-        Minecraft.getInstance().mouseHandler.grabMouse();
+        EnvExecutor.safeRunWhenOn(EnvType.CLIENT, () -> () -> {
+            Minecraft.getInstance().mouseHandler.grabMouse();
+        });
     }
     
     @Override
     @OnlyIn(Dist.CLIENT)
     public void finished(CamRun run) {
         super.finished(run);
-        Minecraft mc = Minecraft.getInstance();
-        if (!mc.player.isCreative())
-            mc.player.getAbilities().flying = false;
+        EnvExecutor.safeRunWhenOn(EnvType.CLIENT, () -> () -> {
+            Minecraft mc = Minecraft.getInstance();
+            if (!mc.player.isCreative())
+                mc.player.getAbilities().flying = false;
+        });
     }
     
     @Override
     @OnlyIn(Dist.CLIENT)
     public Entity getCamera() {
-        return Minecraft.getInstance().player;
+        return EnvExecutor.safeCallWhenOn(EnvType.CLIENT, () -> () -> Minecraft.getInstance().player);
     }
     
     @Override
     @OnlyIn(Dist.CLIENT)
     public void correctTargetPosition(Vec3d vec) {
-        vec.y -= Minecraft.getInstance().player.getEyeHeight();
+        EnvExecutor.safeRunWhenOn(EnvType.CLIENT, () -> () -> {
+            vec.y -= Minecraft.getInstance().player.getEyeHeight();
+        });
     }
     
     @Override
