+++ src/main/java/team/creative/cmdcam/common/scene/mode/OutsideMode.java
@@ -1,5 +1,6 @@
 package team.creative.cmdcam.common.scene.mode;
 
+import net.fabricmc.api.EnvType;
 import net.minecraft.client.Minecraft;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.item.ItemEntity;
@@ -10,6 +11,7 @@ import net.neoforged.api.distmarker.OnlyIn;
 import team.creative.cmdcam.common.math.point.CamPoint;
 import team.creative.cmdcam.common.scene.CamScene;
 import team.creative.cmdcam.common.scene.run.CamRun;
+import team.creative.cmdcam.common.utils.EnvExecutor;
 import team.creative.creativecore.common.util.mc.TickUtils;
 
 public class OutsideMode extends CamMode {
@@ -23,18 +25,22 @@ public class OutsideMode extends CamMode {
     @Override
     @OnlyIn(Dist.CLIENT)
     public void started(CamRun run) {
-        Minecraft mc = Minecraft.getInstance();
-        Vec3 vec = mc.player.getEyePosition(TickUtils.getFrameTime(mc.level));
-        this.camPlayer = new ItemEntity(mc.level, vec.x, vec.y, vec.z, ItemStack.EMPTY);
-        this.camPlayer.setOldPosAndRot();
+        EnvExecutor.safeRunWhenOn(EnvType.CLIENT, () -> () -> {
+            Minecraft mc = Minecraft.getInstance();
+            Vec3 vec = mc.player.getEyePosition(TickUtils.getFrameTime(mc.level));
+            this.camPlayer = new ItemEntity(mc.level, vec.x, vec.y, vec.z, ItemStack.EMPTY);
+            this.camPlayer.setOldPosAndRot();
+        });
     }
     
     @Override
     @OnlyIn(Dist.CLIENT)
     public void finished(CamRun run) {
         super.finished(run);
-        Minecraft mc = Minecraft.getInstance();
-        mc.cameraEntity = mc.player;
+        EnvExecutor.safeRunWhenOn(EnvType.CLIENT, () -> () -> {
+            Minecraft mc = Minecraft.getInstance();
+            mc.cameraEntity = mc.player;
+        });
     }
     
     @Override
@@ -47,7 +53,9 @@ public class OutsideMode extends CamMode {
     @OnlyIn(Dist.CLIENT)
     public void process(CamPoint point) {
         super.process(point);
-        Minecraft.getInstance().cameraEntity = camPlayer;
+        EnvExecutor.safeRunWhenOn(EnvType.CLIENT, () -> () -> {
+            Minecraft.getInstance().cameraEntity = camPlayer;
+        });
     }
     
     @Override
